PR refactoring
